// migration_scripts/release_company_features/scripts/create_users_regroupments/interfaces.ts
export interface DropContactInterface {
    enrichmentDate: string;
    civility?: string;
    first_name?: string;
    last_name?: string;
    full_name?: string;
    email?: Array<{ email?: string; qualification?: string }>;
    phone?: string;
    mobile_phone?: string;
    company?: string;
    website?: string;
    linkedin?: string;
    company_infogreffe?: string;
    siren?: string;
    siret?: string;
    vat?: string;
    nb_employees?: string;
    naf5_code?: string;
    naf5_des?: string;
    siret_address?: string;
    siret_zip?: string;
    siret_city?: string;
    company_linkedin?: string;
    company_turnover?: string;
    company_results?: string;
}

export interface PhoneNumber {
    type: string;
    number: string;
}

export interface Profile {
    _id: mongoose.Schema.Types.ObjectId | string;
    firstName?: string;
    lastName?: string;
    occupation?: string;
    publicIdentifier?: string;
    linkedinId?: number;
    memberId?: string;
    salesMemberId?: string;
    profilePicture?: string;
    influencer?: boolean;
    jobSeeker?: boolean;
    openLink?: boolean;
    premium?: boolean;
    profileUrl?: string;
    region?: string;
    company?: {
        name?: string;
        linkedinUrl?: string;
        website?: string;
    };
    birthday?: {
        month: number;
        day: number;
    };
    address?: string;
    email?: string;
    phoneNumbers?: PhoneNumber[];
    dropContactEnrichment?: DropContactInterface;
}
// ...
export interface UsersRegroupment {
    _id: mongoose.Schema.Types.ObjectId | string;
    users: Array<string>;
    company?: string;
    duplicatedProspects: Array<{ profile: Profile; prospects: Array<{ owner: string; _id: mongoose.Schema.Types.ObjectId | string }> }>;
}